[NFC][wasm-reduce] Split out DestructiveReducer
The Reducer class used to do double duty providing general reducer
utilities and also walking the module to do destructive reduction. Split
the module walking part into a new DestructiveReducer class so the
Reducer class providing general utilities can be reasonably reused by
new reducer classes in the future.
